Move trip planner off MapViewController and lint errors in CI
The type body crossed SwiftLint's error threshold, but CI skipped
lint entirely, so main was green on GitHub and red locally.
